rename repo & pkg
