+++ .github/ISSUE_TEMPLATE/config.yml
@@ -1,5 +1,5 @@
 blank_issues_enabled: false
 contact_links:
   - name: Ask a Question
-    url: https://github.com/Borda/kaggle_brain-tumor-3D/discussions/new
+    url: https://github.com/Borda/kaggle_vol-3D-classify/discussions/new
     about: Ask and answer related questions
