+++ kaggle_volclassif/__init__.py
@@ -4,7 +4,7 @@ __version__ = "0.4.0"
 __docs__ = "Tooling for Kaggle ..."
 __author__ = "Jiri Borovec"
 __author_email__ = "jirka@pytorchlightning.ai"
-__homepage__ = "https://github.com/Borda/kaggle_brain-tumor-3D"
+__homepage__ = "https://github.com/Borda/kaggle_vol-3D-classify"
 __license__ = "MIT"
 
 _PATH_PACKAGE = os.path.realpath(os.path.dirname(__file__))
