+++ tests/test_models.py
@@ -4,8 +4,8 @@ import pytest
 from monai.networks.nets import EfficientNetBN, resnet18
 from pytorch_lightning import seed_everything, Trainer
 
-from kaggle_brain3d.data import BrainScansDM
-from kaggle_brain3d.models import LitBrainMRI, make_submission
+from kaggle_volclassif.data import BrainScansDM
+from kaggle_volclassif.models import LitBrainMRI, make_submission
 from tests.test_data import _generate_synthetic_dataset
 
 _PATH_HERE = os.path.dirname(__file__)
